+++ domain/src/main/java/com/ongveloper/domain/usecase/DeleteScheduleUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteScheduleUseCase @Inject constructor(
     private val scheduleRepository: ScheduleRepository,
 ) {
-    suspend operator fun invoke(id: Long): Result<Long> =
+    suspend operator fun invoke(id: Long): Result<Int> =
         runCatching {
             scheduleRepository.deleteSchedule(id)
         }
